Decompile Higurashi Rei (Mangagamer) using ILSpy 5.0.2 (v4 failed to decompile?)  - Decompiled using ILSpy 5.0.2  - Fixed attributes using our reorder-attributes-v1.0.0  - Fixed DLL imports (now using .net 2.0 among other changes)
Recompiled DLL seems to run game OK

+++ Assets.Scripts.Core.Scene/BlackBars.cs
@@ -13,7 +13,7 @@ namespace Assets.Scripts.Core.Scene
 			{
 				num *= -1;
 			}
-			base.transform.localPosition = new Vector3((float)num, 0f, 0f);
+			base.transform.localPosition = new Vector3(num, 0f, 0f);
 		}
 	}
 }
